Miband and libre improvements merge master

+++ app/src/main/java/com/eveningoutpost/dexdrip/utilitymodels/BgSparklineBuilder.java
@@ -37,8 +37,8 @@ public class BgSparklineBuilder {
     protected BgGraphBuilder bgGraphBuilder;
     protected LineChartView chart;
     protected long endTimeInMillis = new Date().getTime();
-    protected long end = BgGraphBuilder.timestampToFuzzedGraphPos(endTimeInMillis);
-    protected long start = BgGraphBuilder.timestampToFuzzedGraphPos(endTimeInMillis- 60000*180); // 3h
+    protected long end = (long) BgGraphBuilder.timestampToFuzzedGraphPos(endTimeInMillis);
+    protected long start = (long) BgGraphBuilder.timestampToFuzzedGraphPos(endTimeInMillis- 60000*180); // 3h
     protected boolean showLowLine = false;
     protected boolean showHighLine = false;
     protected boolean showAxes = false;
@@ -51,12 +51,12 @@ public class BgSparklineBuilder {
     protected final static int SCALE_TRIGGER = 84;
 
     public BgSparklineBuilder setStart(long start) {
-        this.start = BgGraphBuilder.timestampToFuzzedGraphPos(start);
+        this.start = (long) BgGraphBuilder.timestampToFuzzedGraphPos(start);
         return this;
     }
 
     public BgSparklineBuilder setEnd(long end) {
-        this.end = BgGraphBuilder.timestampToFuzzedGraphPos(end);
+        this.end = (long) BgGraphBuilder.timestampToFuzzedGraphPos(end);
         return this;
     }
